Improvements

+++ src/main/java/com/mauriciotogneri/idlebattle/game/Player.java
@@ -117,7 +117,7 @@ public class Player
     {
         return new PlayerStatus(
                 points,
-                (int) money,
+                money,
                 mineLevel,
                 attackLevel
         );
