+++ src/main/java/com/mauriciotogneri/idlebattle/messages/PlayerStatus.java
@@ -3,11 +3,11 @@ package com.mauriciotogneri.idlebattle.messages;
 public class PlayerStatus
 {
     private final int points;
-    private final int money;
+    private final double money;
     private final int mineLevel;
     private final int attackLevel;
 
-    public PlayerStatus(int points, int money, int mineLevel, int attackLevel)
+    public PlayerStatus(int points, double money, int mineLevel, int attackLevel)
     {
         this.points = points;
         this.money = money;
